lt03wifi: Do proper triple buffering
AOSP keeps 2 surface buffers at all time in case the hwcomposer
misses the time to swap buffers (in cases where it takes 16ms or
less).
Most OEMs modifiy this to 3 to avoid these timing issues, so we do
as well. Keeping 5 buffers available is excessive and probably
drains battery as well, so get rid of it.